remove scripts field from setup.py
The scripts field was pointing to a file in the current directory,
which did not exist.

The scripts field was not needed, as an entry point was defined,
later. The scripts field is intended for actual scripts (executable as
standalone scripts), and you don't take that route, here.

As the intended function was covered by the entry_points field, I
removed the scripts field.