fix tests and refactor logging module
The logging module contained a single submodule with a single function. This was
more efficiently handled as a single function in __init__.py for the loggers
module. Changing this required further changes in tests and the main module code.

Tests were also modified to account for refactoring in an earlier commit.