fix blank database creation when no UniProt search result is returned
The problem being solved was creation of an empty dataframe, where the data values
were not valid for dataframe creation. This was remedied by (i) defining a single
blank dataset (avoids repetition/potential accidental bugs from multiple
occurrences), (ii) changing the datatype for each of the columns so a list of
"NA"s is provided, rather than "NA" strings.

Although the code now runs, it needs to be revised so as not to make tens of
thousands of calls to UniProt for each genome.