update tests to use pytest fixtures
Add licencing and ownership detail to tests. Create conftest.py file to store global test fixtures. Use Monkeypatch to mock call to Entrez to retrieve a scientific name and taxonomy ID [warning: mocking fails becuase of missing __enter__ and __exit__ attributes as result of the call to entrez being contained within an if statement in the original code, which the current structure of the tests does not support. Needto further improve tests].