Update test for get_ncbi_genomes
Rename tests so test target/aim is clearer. Add author and licence details. Add Pytest fixture utilisation, coinciding with the removal of unit-test class structure. Add test to test downloading of GenBank file from NCBI. Add test mocking to mock call to Entrez [Warning! The mocking is still in development and thus two of the tests fail because the mocking is not set up to match the calling method (controlled using 'with') in get_ncbi_genomes].