Update retrieval of rows after EC number search
EC number search using Pandas .str.contains retrieves not a list of the index numbers but the series index numbers and boolean result of the search. Therefore, the approach of retrieving the rows of interest was changed to use df[index_results] to retrieve the rows of interest. Then pandas .duplicated function was added to remove duplicates - although the duplicated function could be called with no arguments to search for duplicated across the entire series content, the duplicated function is passed a single column in the hope this will increase processing speed.